organize Parsing
