+++ src/Data/Morpheus/Parser/Body.hs
@@ -5,29 +5,32 @@ module Data.Morpheus.Parser.Body
   ) where
 
 import           Data.Morpheus.Parser.Arguments                (maybeArguments)
-import           Data.Morpheus.Parser.Internal                 (Parser)
-import           Data.Morpheus.Parser.Primitive                (qualifier, token)
-import           Data.Morpheus.Parser.Terms                    (onType, parseAssignment, spreadLiteral)
+import           Data.Morpheus.Parsing.Internal.Internal       (Parser)
+import           Data.Morpheus.Parsing.Internal.Primitive      (qualifier, token)
+import           Data.Morpheus.Parsing.Internal.Terms          (onType, parseAssignment, spreadLiteral)
 import           Data.Morpheus.Types.Internal.AST.RawSelection (Fragment (..), RawArguments, RawSelection (..),
                                                                 RawSelection' (..), RawSelectionSet, Reference (..))
 import           Data.Text                                     (Text)
-import           Text.Megaparsec                               (sepEndBy, between, getSourcePos, label, many, try, (<|>))
+import           Text.Megaparsec                               (between, getSourcePos, label, many, sepEndBy, try,
+                                                                (<|>))
 import           Text.Megaparsec.Char                          (char, space)
 
 spread :: Parser (Text, RawSelection)
-spread = label "spread" $ do
-  index <- spreadLiteral
-  key' <- token
-  return (key', Spread $ Reference {referenceName = key', referencePosition = index})
+spread =
+  label "spread" $ do
+    index <- spreadLiteral
+    key' <- token
+    return (key', Spread $ Reference {referenceName = key', referencePosition = index})
 
 inlineFragment :: Parser (Text, RawSelection)
-inlineFragment = label "InlineFragment" $ do
-  index <- spreadLiteral
-  type' <- onType
-  fragmentBody <- entries
-  pure
-    ( "INLINE_FRAGMENT"
-    , InlineFragment $ Fragment {fragmentType = type', fragmentSelection = fragmentBody, fragmentPosition = index})
+inlineFragment =
+  label "InlineFragment" $ do
+    index <- spreadLiteral
+    type' <- onType
+    fragmentBody <- entries
+    pure
+      ( "INLINE_FRAGMENT"
+      , InlineFragment $ Fragment {fragmentType = type', fragmentSelection = fragmentBody, fragmentPosition = index})
 
 {-
   accept:
@@ -37,11 +40,12 @@ inlineFragment = label "InlineFragment" $ do
   - field () {...}
 -}
 parseSelectionField :: Parser (Text, RawSelection)
-parseSelectionField = label "SelectionField" $ do
-  (name', position') <- qualifier
-  arguments' <- maybeArguments
-  value' <- body arguments' <|> buildField arguments' position'
-  return (name', value')
+parseSelectionField =
+  label "SelectionField" $ do
+    (name', position') <- qualifier
+    arguments' <- maybeArguments
+    value' <- body arguments' <|> buildField arguments' position'
+    return (name', value')
   where
     buildField arguments' position' =
       pure
@@ -49,25 +53,21 @@ parseSelectionField = label "SelectionField" $ do
          RawSelection' {rawSelectionArguments = arguments', rawSelectionRec = (), rawSelectionPosition = position'})
 
 alias :: Parser (Text, RawSelection)
-alias = label "alias" $ do
-  ((name', position'), selection') <- parseAssignment qualifier parseSelectionField
-  return (name', RawAlias {rawAliasPosition = position', rawAliasSelection = selection'})
+alias =
+  label "alias" $ do
+    ((name', position'), selection') <- parseAssignment qualifier parseSelectionField
+    return (name', RawAlias {rawAliasPosition = position', rawAliasSelection = selection'})
 
 entries :: Parser RawSelectionSet
-entries = label "entries" $
-  between
-    (char '{' *> space)
-    (char '}' *> space)
-    (entry `sepEndBy` many (char ',' *> space))
+entries = label "entries" $ between (char '{' *> space) (char '}' *> space) (entry `sepEndBy` many (char ',' *> space))
   where
-    entry = label "entry" $
-      try inlineFragment <|> try spread <|> try alias <|> parseSelectionField
-
+    entry = label "entry" $ try inlineFragment <|> try spread <|> try alias <|> parseSelectionField
 
 body :: RawArguments -> Parser RawSelection
-body args = label "body" $ do
-  index <- getSourcePos
-  entries' <- entries
-  return
-    (RawSelectionSet $
-     RawSelection' {rawSelectionArguments = args, rawSelectionRec = entries', rawSelectionPosition = index})
+body args =
+  label "body" $ do
+    index <- getSourcePos
+    entries' <- entries
+    return
+      (RawSelectionSet $
+       RawSelection' {rawSelectionArguments = args, rawSelectionRec = entries', rawSelectionPosition = index})
